requirement.txt
aioapns==1.7
amqp==2.5.1
asn1crypto==0.24.0
aspy.yaml==1.3.0
astroid==2.3.2
authy==2.2.3
autopep8==1.4.4
backports.csv==1.0.7
billiard==3.6.1.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
cfgv==1.6.0
chardet==3.0.4
Click==7.0
cryptography==2.7
Cython==0.29.12
defusedxml==0.6.0
Django==2.2.1
django-dynamic-db-router==0.3.0
django-modeltranslation==0.14
django-redis==4.10.0
django-silk==3.0.2
django-tables2==2.1.0
django-uuidfield==0.5.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docopt==0.4.0
et-xmlfile==1.0.1
fcm-django==0.2.21
filelock==3.0.12
geographiclib==1.49
geopy==1.20.0
googletrans==2.4.0
gprof2dot==2017.9.19
h2==3.1.0
hpack==3.0.0
hyperframe==5.2.0
identify==1.4.3
idna==2.8
importlib-metadata==0.19
importlib-resources==1.0.2
isort==4.3.21
jdcal==1.4.1
Jinja2==2.10.1
kombu==4.6.4
lazy-object-proxy==1.4.3
lxml==4.3.3
mailchimp==2.0.9
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
mysqlclient==1.4.2.post1
nodeenv==1.3.3
odfpy==1.4.0
openpyxl==2.6.2
phonenumbers==8.8.4
Pillow==6.0.0
pluggy==0.11.0
pre-commit==1.16.1
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyfcm==1.4.5
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.3
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-dotenv==0.7.1
python-memcached==1.59
pytz==2019.1
PyYAML==5.1
redis==3.3.7
requests==2.22.0
requests-toolbelt==0.9.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
toml==0.10.0
tox==3.12.1
translate==3.5.0
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
virtualenv==16.6.0
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==0.5.1
